FileHunter Mark One
This file hunter mark, saves files into a json format for later parsing inside of a text file, it find files, and directory's and apends their current working directory to the begining of the json data for later parsing. It then print all results to the screen, If results are being printed to the screen then it has already, written to or saved the file and the program can be closed without harm to the file.